feat: evaluate and optuna working
